better setup script

+++ utils/setup/setup_device.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 # USAGE: setup_device.sh [device_id]
+if [[ $UID != 0 ]]; then
+    echo "Please run this script with sudo:"
+    echo "sudo $0 $*"
+    exit 1
+fi
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 sudo rm /etc/hosts
-sudo cp hosts.template  /etc/hosts
+sudo cp $DIR/hosts.template  /etc/hosts
 sudo echo -e "127.0.1.1	        twilight_$1" >> /etc/hosts
 sudo echo -e "twilight_$1" > /etc/hostname
